Rename module holding while loop
